Add leak detector, remove pipeline, split ImageRef god object
- Add atomic counter tracking open ImageRef instances with
  OpenImageRefs() and AssertNoLeaks(t) test helpers
- Fix Black() to use newImageRef() so it participates in leak tracking
- Remove vips/pipeline/ package (thin 1:1 wrapper over ImageRef,
  covered ~40 of 139 methods, creating dual-surface confusion)
- Split image.go (2607 lines) into 8 focused files by concern:
  image.go (core/lifecycle), image_metadata.go, image_transform.go,
  image_color.go, image_pixel.go, image_composite.go,
  image_export.go, image_icc.go
- Zero API changes, all tests pass
